Pipeline: sampler as a plugin/action
