+++ pydatview/GUITools.py
@@ -22,7 +22,15 @@ class GUIToolPanel(wx.Panel):
         super(GUIToolPanel,self).__init__(parent)
         self.parent   = parent
 
-    def destroy(self,event=None):
+    def destroyData(self):
+        if hasattr(self, 'action'):
+            if self.action is not None:
+                # cleanup action calls
+                self.action.guiEditorObj = None
+                self.action = None
+
+    def destroy(self,event=None): # TODO rename to something less close to Destroy
+        self.destroyData()
         self.parent.removeTools()
 
     def getBtBitmap(self,par,label,Type=None,callback=None,bitmap=False):
@@ -443,245 +451,6 @@ To filter perform the following step:
           Currently done for all tables.
 """)
 
-
-# --------------------------------------------------------------------------------}
-# --- Resample
-# --------------------------------------------------------------------------------{
-class ResampleToolPanel(GUIToolPanel):
-    def __init__(self, parent):
-        super(ResampleToolPanel,self).__init__(parent)
-
-        # --- Data from other modules
-        from pydatview.tools.signal_analysis import SAMPLERS
-        self.parent = parent # parent is GUIPlotPanel
-        self._SAMPLERS=SAMPLERS 
-        # Setting default states to parent
-        if 'Sampler' not in self.parent.plotDataOptions.keys():
-            self.parent.plotDataOptions['Sampler']=None
-        self._applied = type(self.parent.plotDataOptions['Sampler'])==dict
-
-
-        # --- GUI elements
-        self.btClose    = self.getBtBitmap(self, 'Close','close', self.destroy)
-        self.btAdd      = self.getBtBitmap(self, 'Add','add'  , self.onAdd)
-        self.btHelp     = self.getBtBitmap(self, 'Help','help', self.onHelp)
-        self.btClear    = self.getBtBitmap(self, 'Clear Plot','sun', self.onClear)
-        self.btPlot     = self.getBtBitmap(self, 'Plot' ,'chart'  , self.onPlot)
-        self.btApply    = self.getToggleBtBitmap(self,'Apply','cloud',self.onToggleApply)
-
-        #self.lb         = wx.StaticText( self, -1, """ Click help """)
-        self.cbTabs     = wx.ComboBox(self, -1, choices=[], style=wx.CB_READONLY)
-        self.cbMethods  = wx.ComboBox(self, -1, choices=[s['name'] for s in self._SAMPLERS], style=wx.CB_READONLY)
-
-        self.lbNewX   = wx.StaticText(self, -1, 'New x:  ')
-        self.textNewX = wx.TextCtrl(self, wx.ID_ANY, '', style      = wx.TE_PROCESS_ENTER)
-        self.textOldX = wx.TextCtrl(self, wx.ID_ANY|wx.TE_READONLY)
-        self.textOldX.Enable(False)
-
-        # --- Layout
-        btSizer  = wx.FlexGridSizer(rows=3, cols=2, hgap=2, vgap=0)
-        btSizer.Add(self.btClose                , 0, flag = wx.ALL|wx.EXPAND, border = 1)
-        btSizer.Add(self.btClear                , 0, flag = wx.ALL|wx.EXPAND, border = 1)
-        btSizer.Add(self.btAdd                  , 0, flag = wx.ALL|wx.EXPAND, border = 1)
-        btSizer.Add(self.btPlot                 , 0, flag = wx.ALL|wx.EXPAND, border = 1)
-        btSizer.Add(self.btHelp                 , 0, flag = wx.ALL|wx.EXPAND, border = 1)
-        btSizer.Add(self.btApply                , 0, flag = wx.ALL|wx.EXPAND, border = 1)
-
-        msizer  = wx.FlexGridSizer(rows=2, cols=4, hgap=2, vgap=0)
-        msizer.Add(wx.StaticText(self, -1, 'Table:')    , 0, wx.ALIGN_LEFT|wx.ALIGN_CENTER_VERTICAL|wx.TOP|wx.BOTTOM, 1)
-        msizer.Add(self.cbTabs                          , 0, wx.ALIGN_LEFT|wx.ALIGN_CENTER_VERTICAL|wx.TOP|wx.BOTTOM, 1)
-        msizer.Add(wx.StaticText(self, -1, 'Current x:          '), 0, wx.ALIGN_LEFT|wx.ALIGN_CENTER_VERTICAL|wx.TOP|wx.BOTTOM, 1)
-        msizer.Add(self.textOldX                        , 1, wx.ALIGN_LEFT|wx.ALIGN_CENTER_VERTICAL|wx.TOP|wx.BOTTOM|wx.EXPAND, 1)
-        msizer.Add(wx.StaticText(self, -1, 'Method:')   , 0, wx.ALIGN_LEFT|wx.ALIGN_CENTER_VERTICAL|wx.TOP|wx.BOTTOM, 1)
-        msizer.Add(self.cbMethods                       , 0, wx.ALIGN_LEFT|wx.ALIGN_CENTER_VERTICAL|wx.TOP|wx.BOTTOM, 1)
-        msizer.Add(self.lbNewX                          , 0, wx.ALIGN_LEFT|wx.ALIGN_CENTER_VERTICAL|wx.TOP|wx.BOTTOM, 1)
-        msizer.Add(self.textNewX                        , 1, wx.ALIGN_LEFT|wx.ALIGN_CENTER_VERTICAL|wx.TOP|wx.BOTTOM|wx.EXPAND, 1)
-        msizer.AddGrowableCol(3,1)
-
-        self.sizer = wx.BoxSizer(wx.HORIZONTAL)
-        self.sizer.Add(btSizer  ,0, flag = wx.LEFT           ,border = 5)
-        self.sizer.Add(msizer   ,1, flag = wx.LEFT|wx.EXPAND ,border = TOOL_BORDER)
-        self.SetSizer(self.sizer)
-
-        # --- Events
-        self.cbTabs.Bind   (wx.EVT_COMBOBOX, self.onTabChange)
-        self.cbMethods.Bind(wx.EVT_COMBOBOX, self.onMethodChange)
-        self.textNewX.Bind(wx.EVT_TEXT_ENTER,self.onParamChange)
-
-        # --- Init triggers
-        self.cbMethods.SetSelection(3)
-        self.onMethodChange(init=True)
-        self.onToggleApply(init=True)
-        self.updateTabList()
-        self.textNewX.SetValue('2')
-
-    def setCurrentX(self, x=None):
-        if x is None:
-            x= self.parent.plotData[0].x
-        if len(x)<50:
-            s=np.array2string(x, separator=', ')
-        else:
-            s =np.array2string(x[[0,1,2,3]], separator=', ') 
-            s+=', ...,  '
-            s+=np.array2string(x[[-3,-2,-1]], separator=', ') 
-        s=s.replace('[','').replace(']','').replace(' ','').replace(',',', ')
-
-        self.textOldX.SetValue(s)
-
-    def onMethodChange(self, event=None, init=True):
-        """ Select the method, but does not applied it to the plotData 
-            User data and option is unchanged
-            But if the user already has some options, they are used
-        """
-        iOpt = self.cbMethods.GetSelection()
-        opt = self._SAMPLERS[iOpt]
-        self.lbNewX.SetLabel(opt['paramName']+':')
-
-        parentOpt=self.parent.plotDataOptions['Sampler']
-        # Value
-        if len(self.textNewX.Value)==0:
-            if type(parentOpt)==dict:
-                self.textNewX.SetValue(str(parentOpt['param'])[1:-1])
-            else:
-                self.textNewX.SetValue(str(opt['param'])[2:-2])
-        self.onParamChange()
-
-    def onParamChange(self, event=None):
-        if self._applied:
-            self.parent.plotDataOptions['Sampler'] =self._GUI2Data()
-            self.parent.load_and_draw() # Data will change
-            self.setCurrentX()
-
-    def _GUI2Data(self):
-        iOpt = self.cbMethods.GetSelection()
-        opt = self._SAMPLERS[iOpt].copy()
-        s= self.textNewX.Value.strip().replace('[','').replace(']','')
-        if len(s)>0:
-            if s.find(','):
-                opt['param']=np.array(s.split(',')).astype(float)
-            else:
-                opt['param']=np.array(s.split('')).astype(float)
-        return opt
-
-    def onToggleApply(self, event=None, init=False):
-        """
-        apply sampler based on GUI Data
-        """
-        parentFilt=self.parent.plotDataOptions['Sampler']
-        if not init:
-            self._applied = not self._applied
-
-        if self._applied:
-            self.parent.plotDataOptions['Sampler'] =self._GUI2Data()
-            #print('Apply', self.parent.plotDataOptions['Sampler'])
-            #self.lbInfo.SetLabel(
-            #        'Sampler is now applied on the fly. Change parameter live. Click "Clear" to stop. '
-            #        )
-            self.btPlot.Enable(False)
-            self.btClear.Enable(False)
-            self.btApply.SetLabel(CHAR['sun']+' Clear')
-        else:
-            self.parent.plotDataOptions['Sampler'] = None
-            #self.lbInfo.SetLabel(
-            #        'Click on "Apply" to set filter on the fly for all plots. '+
-            #        'Click on "Plot" to try a filter on the current plot.'
-            #        )
-            self.btPlot.Enable(True)
-            self.btClear.Enable(True)
-            self.btApply.SetLabel(CHAR['cloud']+' Apply')
-
-        if not init:
-            self.parent.load_and_draw() # Data will change
-        self.setCurrentX()
-
-    def onAdd(self,event=None):
-        iSel         = self.cbTabs.GetSelection()
-        tabList      = self.parent.selPanel.tabList
-        icol, colname = self.parent.selPanel.xCol
-        opt = self._GUI2Data()
-        errors=[]
-        if iSel==0:
-            dfs, names, errors = tabList.applyResampling(icol, opt, bAdd=True)
-            self.parent.addTables(dfs,names,bAdd=True)
-        else:
-            df, name = tabList.get(iSel-1).applyResampling(icol, opt, bAdd=True)
-            self.parent.addTables([df],[name], bAdd=True)
-        self.updateTabList()
-
-        if len(errors)>0:
-            raise Exception('Error: The resampling failed on some tables:\n\n'+'\n'.join(errors))
-
-        # We stop applying
-        self.onToggleApply()
-
-    def onPlot(self,event=None):
-        from pydatview.tools.signal_analysis import applySampler
-        if len(self.parent.plotData)!=1:
-            Error(self,'Plotting only works for a single plot. Plot less data.')
-            return
-        opts=self._GUI2Data()
-        PD = self.parent.plotData[0]
-        x_new, y_new = applySampler(PD.x0, PD.y0, opts)
-        ax = self.parent.fig.axes[0]
-
-        PD_new = PlotData()
-        PD_new.fromXY(x_new, y_new)
-        self.parent.transformPlotData(PD_new)
-        ax.plot(PD_new.x, PD_new.y, '-')
-        self.setCurrentX(x_new)
-
-        self.parent.canvas.draw()
-
-    def onClear(self,event=None):
-        self.parent.load_and_draw() # Data will change
-        # Update Current X
-        self.setCurrentX()
-        # Update Table list
-        self.updateTabList()
-
-
-    def onTabChange(self,event=None):
-        #tabList = self.parent.selPanel.tabList
-        #iSel=self.cbTabs.GetSelection()
-        pass
-
-    def updateTabList(self,event=None):
-        tabList = self.parent.selPanel.tabList
-        tabListNames = ['All opened tables']+tabList.getDisplayTabNames()
-        try:
-            iSel=np.max([np.min([self.cbTabs.GetSelection(),len(tabListNames)]),0])
-            self.cbTabs.Clear()
-            [self.cbTabs.Append(tn) for tn in tabListNames]
-            self.cbTabs.SetSelection(iSel)
-        except RuntimeError:
-            pass
-
-    def onHelp(self,event=None):
-        Info(self,"""Resampling.
-
-The resampling operation changes the "x" values of a table/plot and 
-adapt the "y" values accordingly.
-
-To resample perform the following step:
-
-- Chose a resampling method:
-   - replace: specify all the new x-values
-   - insert : insert a list of x values to the existing ones
-   - delete : delete a list of x values from the existing ones
-   - every-n : use every n values 
-   - time-based: downsample using sample averaging or upsample using
-                 linear interpolation, x-axis must already be in seconds
-   - delta x : specify a delta for uniform spacing of x values
-
-- Specify the x values as a space or comma separated list
-
-- Click on one of the following buttons:
-   - Plot: will display the resampled data on the figure
-   - Apply: will perform the resampling on the fly for all new plots
-   - Add: will create new table(s) with resampled values for all 
-          signals. This process might take some time.
-          Select a table or choose all (default)
-""")
 
 
 
@@ -1169,7 +938,6 @@ TOOLS={
  'LogDec':    LogDecToolPanel,
  'Outlier':   OutlierToolPanel,
  'Filter':    FilterToolPanel,
- 'Resample':  ResampleToolPanel,
  'Mask':      MaskToolPanel, 
  'FASTRadialAverage': RadialToolPanel,
  'CurveFitting': CurveFitToolPanel,
